restructure validation utilities
