+++ src/validators/Validator.ts
@@ -1,6 +1,6 @@
 import { Request } from 'express';
 
-import { ValidationResult } from './utils/ValidationResult';
+import { ValidationResult } from '../utils/validation/ValidationResult';
 
 export interface Validator {
   validate(request: Request): Promise<ValidationResult>;
